+++ ChangeLog.md
@@ -1,3 +1,7 @@
+###### v0.10.14 (2020-10-07):
+ * Prepared structure for measure information modal.
+ * Added brief description of the summary table in the main panel.
+
 ###### v0.10.13 (2020-10-07):
  * Renamed `measure_info` modal to `k_param_info` to separate these two informations.
  * Changed cursor for the info icons.